+++ app/services/AppConfigurationService.scala
@@ -116,8 +116,20 @@ object AppConfiguration {
   /** Set the google analytics code */
   def setGoogleAnalytics(gacode: String) = appConfig.setProperty("google.analytics", gacode)
 
-  /** Get the welcome message */
+  /** Get the google analytics code */
   def getGoogleAnalytics: String = appConfig.getProperty("google.analytics", "")
+
+  // ----------------------------------------------------------------------
+
+  /** Set the Amplitude clickstream/analytics configuration */
+  def setAmplitudeApiKey(ampApiKey: String) = {
+    appConfig.setProperty("amplitude.apikey", ampApiKey)
+  }
+
+  /** Get the Amplitude clickstream/analytics configuration */
+  def getAmplitudeApiKey: String = {
+    appConfig.getProperty("amplitude.apikey", "")
+  }
 
   // ----------------------------------------------------------------------