+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+- Added support for Amplitude clickstream tracking. See Admin -> Customize to configure Amplitude apikey.
+
+
 ## 1.12.1 - 2020-11-05
 
 ### Fixed